Break the soon-to-be-only visitor out
